requirements.txt
fastapi==0.118.3
passlib==1.7.4
psycopg2_binary==2.9.10
pydantic==2.12.0
pydantic_settings==2.11.0
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.34
sqlmodel==0.0.27
